requirements.txt
gradio==3.47.1
gradio_client==0.6.0
python-dotenv==1.0.0
transformers==4.33.3
langchain==0.0.325
requests==2.31.0
requests-oauthlib==1.3.1